DEVS-48_avatar refactor
